+++ packages/cspell-lib/src/Settings/ImportError.ts
@@ -1,5 +1,9 @@
+import { isError } from '../util/errors';
+
 export class ImportError extends Error {
-    constructor(msg: string, readonly cause?: Error) {
+    readonly cause?: Error;
+    constructor(msg: string, cause?: Error | unknown) {
         super(msg);
+        this.cause = isError(cause) ? cause : undefined;
     }
 }