+++ packages/cspell-trie-lib/src/lib/suggestCollector.ts
@@ -3,21 +3,21 @@ import { JOIN_SEPARATOR, WORD_SEPARATOR } from './walker';
 
 const defaultMaxNumberSuggestions = 10;
 
-const baseCost = 100;
-const maxNumChanges = 5;
-const maxCostScale = 0.5;
+const BASE_COST = 100;
+const MAX_NUM_CHANGES = 5;
+const MAX_COST_SCALE = 0.5;
 // max allowed cost scale should be a bit over 50% to allow for suggestions to short words, but not too high to have too many suggestions.
-const maxAllowedCostScale = 1.03 * maxCostScale;
+const MAX_ALLOWED_COST_SCALE = 1.03 * MAX_COST_SCALE;
 
 const collator = new Intl.Collator();
 
 const regexSeparator = new RegExp(`[${regexQuote(JOIN_SEPARATOR + WORD_SEPARATOR)}]`, 'g');
 
 const wordLengthCost = [0, 50, 25, 5, 0];
-const extraWordsCost = 5;
+const EXTRA_WORD_COST = 5;
 
 /** time in ms */
-const defaultCollectorTimeout = 1000;
+const DEFAULT_COLLECTOR_TIMEOUT = 1000;
 
 export type Cost = number;
 export type MaxCost = Cost;
@@ -58,6 +58,8 @@ export function compSuggestionResults(a: SuggestionResult, b: SuggestionResult):
     return a.cost - b.cost || a.word.length - b.word.length || collator.compare(a.word, b.word);
 }
 
+export type FilterWordFn = (word: string, cost: number) => boolean;
+
 export interface SuggestionCollector {
     /**
      * Collection suggestions from a SuggestionIterator
@@ -69,10 +71,10 @@ export interface SuggestionCollector {
      * r = yield(suggestion);
      * if (r === collector.symbolStopProcessing) // ...stop generating suggestions.
      */
-    collect: (src: SuggestionGenerator, timeout?: number) => void;
+    collect: (src: SuggestionGenerator, timeout?: number, filter?: FilterWordFn) => void;
     add: (suggestion: SuggestionResult) => SuggestionCollector;
     readonly suggestions: SuggestionResult[];
-    readonly maxNumChanges: number;
+    readonly changeLimit: number;
     readonly maxCost: number;
     readonly word: string;
     readonly maxNumSuggestions: number;
@@ -87,44 +89,64 @@ export interface SuggestionCollector {
 export interface SuggestionCollectorOptions {
     /**
      * number of best matching suggestions.
+     * @default 10
      */
     numSuggestions: number;
 
     /**
      * An optional filter function that can be used to limit remove unwanted suggestions.
      * I.E. to remove forbidden terms.
+     * @default () => true
      */
-    filter?: (word: string, cost: number) => boolean;
+    filter?: FilterWordFn;
 
     /**
      * The number of letters that can be changed when looking for a match
+     * @default 5
      */
     changeLimit: number | undefined;
 
     /**
      * Include suggestions with tied cost even if the number is greater than `numSuggestions`.
+     * @default true
      */
     includeTies?: boolean;
 
     /**
      * specify if case / accents should be ignored when looking for suggestions.
+     * @default true
      */
     ignoreCase: boolean | undefined;
+
+    /**
+     * the total amount of time to allow for suggestions.
+     * @default 1000
+     */
+    timeout?: number | undefined;
 }
 
 export const defaultSuggestionCollectorOptions: SuggestionCollectorOptions = {
     numSuggestions: defaultMaxNumberSuggestions,
     filter: () => true,
-    changeLimit: maxNumChanges,
-    includeTies: true,
+    changeLimit: MAX_NUM_CHANGES,
+    includeTies: false,
     ignoreCase: true,
+    timeout: DEFAULT_COLLECTOR_TIMEOUT,
 };
 
 export function suggestionCollector(wordToMatch: string, options: SuggestionCollectorOptions): SuggestionCollector {
-    const { filter = () => true, changeLimit = maxNumChanges, includeTies = false, ignoreCase = true } = options;
+    const {
+        filter = () => true,
+        changeLimit = MAX_NUM_CHANGES,
+        includeTies = false,
+        ignoreCase = true,
+        timeout = DEFAULT_COLLECTOR_TIMEOUT,
+    } = options;
     const numSuggestions = Math.max(options.numSuggestions, 0) || 0;
     const sugs = new Map<string, SuggestionResult>();
-    let maxCost: number = baseCost * Math.min(wordToMatch.length * maxAllowedCostScale, changeLimit);
+    let maxCost: number = BASE_COST * Math.min(wordToMatch.length * MAX_ALLOWED_COST_SCALE, changeLimit);
+
+    let timeRemaining = timeout;
 
     function dropMax() {
         if (sugs.size < 2) {
@@ -146,7 +168,7 @@ export function suggestionCollector(wordToMatch: string, options: SuggestionColl
         const words = sug.word.split(regexSeparator);
         const extraCost =
             words.map((w) => wordLengthCost[w.length] || 0).reduce((a, b) => a + b, 0) +
-            (words.length - 1) * extraWordsCost;
+            (words.length - 1) * EXTRA_WORD_COST;
         return { word: sug.word, cost: sug.cost + extraCost };
     }
 
@@ -175,8 +197,12 @@ export function suggestionCollector(wordToMatch: string, options: SuggestionColl
      * @param src - the SuggestionIterator used to generate suggestions.
      * @param timeout - the amount of time in milliseconds to allow for suggestions.
      */
-    function collect(src: SuggestionGenerator, timeout = defaultCollectorTimeout) {
+    function collect(src: SuggestionGenerator, timeout?: number, filter?: FilterWordFn) {
         let stop: false | symbol = false;
+        timeout = timeout ?? timeRemaining;
+        timeout = Math.min(timeout, timeRemaining);
+        if (timeout < 0) return;
+
         const timer = createTimer();
 
         let ir: IteratorResult<SuggestionResult | Progress | undefined>;
@@ -187,11 +213,15 @@ export function suggestionCollector(wordToMatch: string, options: SuggestionColl
             const { value } = ir;
             if (!value) continue;
             if (isSuggestionResult(value)) {
-                collectSuggestion(value);
+                if (!filter || filter(value.word, value.cost)) {
+                    collectSuggestion(value);
+                }
                 continue;
             }
             handleProgress(value);
         }
+
+        timeRemaining -= timer.elapsed();
     }
 
     function suggestions() {
@@ -220,7 +250,7 @@ export function suggestionCollector(wordToMatch: string, options: SuggestionColl
         get maxNumSuggestions() {
             return numSuggestions;
         },
-        get maxNumChanges() {
+        get changeLimit() {
             return changeLimit;
         },
         includesTies: includeTies,