+++ packages/cspell-trie-lib/src/lib/walker.test.ts
@@ -1,5 +1,6 @@
 import { walker, YieldResult, hintedWalker } from './walker';
 import { orderTrie, createTriFromList } from './util';
+import { parseLinesToDictionary } from './SimpleDictionaryParser';
 
 describe('Validate Util Functions', () => {
     test('Tests Walker', () => {
@@ -24,7 +25,7 @@ describe('Validate Util Functions', () => {
         orderTrie(root);
         // cspell:ignore joty
         // prefer letters j, o, t, y before the others.
-        const i = hintedWalker(root, undefined, 'joty');
+        const i = hintedWalker(root, false, 'joty', undefined);
         let goDeeper = true;
         let ir: IteratorResult<YieldResult>;
         const result: string[] = [];
@@ -36,6 +37,40 @@ describe('Validate Util Functions', () => {
             goDeeper = text.length < 4;
         }
         expect(result).toEqual('joy jowl talk lift walk'.split(' '));
+    });
+
+    test('Hinted Walker compounds', () => {
+        const dict = ['A*', '+a*', '*b*', '+c'];
+        const trie = parseLinesToDictionary(dict, { stripCaseAndAccents: true });
+        const i = hintedWalker(trie.root, false, 'a', undefined);
+        let goDeeper = true;
+        let ir: IteratorResult<YieldResult>;
+        const result: string[] = [];
+        while (!(ir = i.next({ goDeeper })).done) {
+            const { text, node } = ir.value;
+            if (node.f) {
+                result.push(text);
+            }
+            goDeeper = text.length < 2;
+        }
+        expect(result).toEqual(['A', 'Aa', 'Ab', 'Ac', 'b', 'ba', 'bb', 'bc']);
+    });
+
+    test('Hinted Walker compounds ignoreCase', () => {
+        const dict = ['A*', '+a*', '*b*', '+c'];
+        const trie = parseLinesToDictionary(dict, { stripCaseAndAccents: true });
+        const i = hintedWalker(trie.root, true, 'a', undefined);
+        let goDeeper = true;
+        let ir: IteratorResult<YieldResult>;
+        const result: string[] = [];
+        while (!(ir = i.next({ goDeeper })).done) {
+            const { text, node } = ir.value;
+            if (node.f) {
+                result.push(text);
+            }
+            goDeeper = text.length < 2;
+        }
+        expect(result).toEqual(['A', 'Aa', 'Ab', 'Ac', 'b', 'ba', 'bb', 'bc', 'a', 'aa', 'ab', 'ac']);
     });
 });